Update stable-diffusion-pretraining to Ray 2.54.1
- Update BUILD.yaml image to anyscale/ray:2.54.1-py311-cu128
- Extract SDTransformer and SDLatentEncoder to sd_data_transforms.py
  to fix GenericModule pickle error with newer PyTorch
- Fix ray.data.set_progress_bars -> ctx.enable_progress_bars (Ray 2.54.0)
- Fix shuffle="file" -> shuffle="files" (Ray 2.54.0)
- Fix get_trial_name() -> get_experiment_name() (Ray Train V2)
- Switch model from stabilityai/stable-diffusion-2-base to
  CompVis/stable-diffusion-v1-4 (public model)
- Update cross_attention_dim from 1024 to 768 for SD v1
- Create tests/stable-diffusion-pretraining/tests.sh

Made-with: Cursor
