+++ BUILD.yaml
@@ -320,16 +320,14 @@
 - name: stable-diffusion-pretraining
   dir: templates/stable-diffusion-pretraining
   cluster_env:
-    byod:
-      docker_image: us-docker.pkg.dev/anyscale-workspace-templates/workspace-templates/stable-diffusion-pretraining:ray-2.12-full
-      ray_version: 2.12.0
+    image_uri: anyscale/ray:2.54.1-py311-cu128
   compute_config:
     AWS: configs/stable-diffusion-pretraining/aws.yaml
     GCP: configs/stable-diffusion-pretraining/gce.yaml
   test:
     tests_path: tests/stable-diffusion-pretraining/
     command: bash tests.sh
-    timeout_in_sec: 900
+    timeout_in_sec: 2400
 
 # owner: @mattd
 - name: parallel-experiments
