+++ templates/stable-diffusion-pretraining/scripts/end_to_end.py
@@ -1,4 +1,4 @@
-"""End-to-end online preprocessing and model training script for Stable Diffusion v2.
+"""End-to-end online preprocessing and model training script for Stable Diffusion.
 
 The script performs the following steps:
 1. Load images and text captions from a remote storage system using the read_data function.
@@ -8,16 +8,14 @@ The script performs the following steps:
 5. Run the scalable training procedure with Ray Train using the train entry point.
 """
 
-import gc
-import io
 import logging
-import math
 import os
 import re
 import shutil
 import tempfile
 from contextlib import nullcontext
 from functools import partial
+from pathlib import Path
 from typing import Literal, Optional, cast, ContextManager, Mapping, Union, Any
 
 import lightning.pytorch as pl  # type: ignore
@@ -27,14 +25,11 @@ import pyarrow.fs
 import ray.train
 import torch
 import torch.nn.functional as F
-import torchvision  # type: ignore
 import typer
 from diffusers import DDPMScheduler, UNet2DConditionModel
-from diffusers.models import AutoencoderKL
 from lightning.pytorch import seed_everything  # type: ignore
 from lightning.pytorch.callbacks import LearningRateMonitor  # type: ignore
 from lightning.pytorch.utilities.types import OptimizerLRScheduler  # type: ignore
-from PIL import Image
 from ray.train import Checkpoint, FailureConfig, RunConfig, ScalingConfig
 from ray.train.lightning import RayDDPStrategy, RayFSDPStrategy, RayLightningEnvironment
 from ray.train.torch import TorchTrainer, get_device
@@ -52,7 +47,17 @@ except ImportError:
     AZURE_IDENTITY_AVAILABLE = False
 from torch.distributed.fsdp import BackwardPrefetch
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
-from transformers import CLIPTextModel, CLIPTokenizer, PretrainedConfig, get_linear_schedule_with_warmup
+from transformers import PretrainedConfig, get_linear_schedule_with_warmup
+
+import sys
+sys.path.insert(0, str(Path(__file__).parent.parent))
+from sd_data_transforms import (
+    SDTransformer,
+    SDLatentEncoder,
+    CAPTION_LATENTS_KEY,
+    IMAGE_LATENTS_256_KEY,
+    IMAGE_LATENTS_512_KEY,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -66,11 +71,6 @@ from abfss_utils import (
     upload_to_abfss,
     get_local_storage_path,
 )
-
-### Constants ###
-CAPTION_LATENTS_KEY = "caption_latents"
-IMAGE_LATENTS_256_KEY = "latents_256_bytes"
-IMAGE_LATENTS_512_KEY = "latents_512_bytes"
 
 
 ########################
@@ -111,245 +111,6 @@ def read_data(
     if limit:
         ds = ds.limit(limit)
     return ds
-
-
-##########################
-# Step 2: Transformation #
-##########################
-
-#### Utils ####
-class LargestCenterSquare:
-    """Largest center square crop for images."""
-
-    def __init__(self, size: int) -> None:
-        self.size = size
-
-    def __call__(self, img: Image.Image) -> Image.Image:
-        """Crop the largest center square from an image."""
-        # First, resize the image such that the smallest side is self.size while preserving aspect ratio.
-        img = torchvision.transforms.functional.resize(
-            img=img,
-            size=self.size,
-        )
-
-        # Then take a center crop to a square.
-        w, h = img.size
-        c_top = (h - self.size) // 2
-        c_left = (w - self.size) // 2
-        img = torchvision.transforms.functional.crop(
-            img=img,
-            top=c_top,
-            left=c_left,
-            height=self.size,
-            width=self.size,
-        )
-        return img
-
-#### Transformer ####
-class SDTransformer:
-    """Image and text transforms."""
-
-    def __init__(
-        self,
-        resolution: int,
-        model_name: str = "stabilityai/stable-diffusion-2-base",
-    ) -> None:
-        # Image transforms.
-        self.resolution = resolution
-        normalize = torchvision.transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
-        self.crop = LargestCenterSquare(resolution)
-        self.transforms = torchvision.transforms.Compose(
-            [torchvision.transforms.ToTensor(), normalize]
-        )
-        # Text tokenizer.
-        self.text_tokenizer = CLIPTokenizer.from_pretrained(
-            model_name, subfolder="tokenizer"
-        )
-
-    def image_transform(self, image: Image.Image) -> np.ndarray:
-        """Transform image to a square-sized normalized tensor."""
-        image = self.crop(image)
-        out = self.transforms(image)
-        return convert_tensor_to_array(out)
-
-    def text_tokenize(self, text: str) -> np.ndarray:
-        """Tokenize text using the CLIP tokenizer into a fixed-length sequence."""
-        return self.text_tokenizer(
-            text,
-            padding="max_length",
-            max_length=self.text_tokenizer.model_max_length,
-            truncation=True,
-            return_tensors="np",
-        )["input_ids"][0]
-
-    def __call__(self, batch: dict[str, np.ndarray]) -> dict[str, np.ndarray]:
-        """Transform images and text captions."""
-        final_batch: dict[str, list] = {
-            "hash": [],
-            "caption_ids": [],
-            f"image_{self.resolution}": [],
-        }
-        logger.info("Running image and text transforms...")
-        for hash_, jpg, width, height, caption in zip(
-            batch["hash"],
-            batch["jpg"],
-            batch["width"],
-            batch["height"],
-            batch["caption"],
-        ):
-            if jpg == b"":
-                logger.info(f"Skipping hash {hash_} due to empty image.")
-                continue
-            if math.isnan(width) or width == 0:
-                logger.info(f"Skipping hash {hash_} due to invalid width.")
-                continue
-            if math.isnan(height) or height == 0:
-                logger.info(f"Skipping hash {hash_} due to invalid height.")
-                continue
-
-            try:
-                image = Image.open(io.BytesIO(jpg))
-            except IOError:
-                logger.info(
-                    f"Skipping hash {hash_} due to invalid image.", exc_info=False
-                )
-                continue
-
-            if image.mode != "RGB":
-                logger.info(f"Converting image to RGB.")
-                image = image.convert("RGB")
-
-            image_arr = self.image_transform(image)
-            caption_ids = self.text_tokenize(caption)
-
-            final_batch["hash"].append(hash_)
-            final_batch[f"image_{self.resolution}"].append(image_arr)
-            final_batch["caption_ids"].append(caption_ids)
-
-        logger.info("Finished executing image and text transforms.")
-        return {k: np.array(v) for k, v in final_batch.items()}
-
-
-####################
-# Step 3: Encoding #
-####################
-
-#### Utils ####
-def convert_tensor_to_array(tensor: torch.Tensor, dtype=np.float32) -> np.ndarray:
-    """Convert a torch tensor to a numpy array."""
-    array = tensor.detach().cpu().numpy()
-    return array.astype(dtype)
-
-
-def supports_float16(device: torch.device) -> bool:
-    """Check if the device supports float16."""
-    if device == torch.device("cuda"):
-        properties = torch.cuda.get_device_properties(device)
-        return properties.major >= 7  # Volta or newer
-    else:
-        return False
-
-
-def resolve_device() -> torch.device:
-    """Resolve to the first available device: CUDA, MPS, or CPU."""
-    if torch.cuda.is_available():
-        return torch.device("cuda")
-    if torch.backends.mps.is_available():
-        return torch.device("mps")
-    return torch.device("cpu")
-
-
-#### Encoder ####
-class SDLatentEncoder:
-    """Latent encoder to encode images and text."""
-
-    def __init__(
-        self,
-        resolution: int = 256,
-        device: Optional[str] = None,
-        model_name: str = "stabilityai/stable-diffusion-2-base",
-    ) -> None:
-        self.device = torch.device(device) if device else resolve_device()
-        self.resolution = resolution
-
-        # Image and text encoders.
-        self.vae = AutoencoderKL.from_pretrained(
-            model_name,
-            subfolder="vae",
-            torch_dtype=(
-                torch.float16 if supports_float16(self.device) else torch.float32
-            ),
-        )
-        self.text_encoder = CLIPTextModel.from_pretrained(
-            model_name,
-            subfolder="text_encoder",
-            torch_dtype=(
-                torch.float16 if supports_float16(self.device) else torch.float32
-            ),
-        )
-
-        # Move the encoders to device.
-        self.vae = self.vae.to(self.device)
-        self.text_encoder = self.text_encoder.to(self.device)
-
-        logger.info("Initialized SDLatentEncoder.")
-        logger.info(f"Device: {self.device}")
-        logger.info(f"Resolution: {self.resolution}")
-
-    @property
-    def image_latents_key(self) -> str:
-        """Return the key for image latents based on resolution."""
-        if self.resolution == 256:
-            return IMAGE_LATENTS_256_KEY
-        elif self.resolution == 512:
-            return IMAGE_LATENTS_512_KEY
-        else:
-            raise ValueError(f"Unsupported resolution: {self.resolution}")
-
-    def encode_images(self, images: np.ndarray) -> np.ndarray:
-        """Encode images into a latent space."""
-        logger.info(f"Encoding images with shape: {images.shape}")
-        input_images = torch.tensor(images, device=self.device)
-
-        if supports_float16(self.device):
-            latent_dist = self.vae.encode(input_images.half())["latent_dist"]
-        else:
-            logger.info("Running VAE.encode")
-            latent_dist = self.vae.encode(input_images)["latent_dist"]
-
-        image_latents = latent_dist.sample() * 0.18215
-        logger.info(f"Image latents shape: {image_latents.shape}")
-        return convert_tensor_to_array(image_latents)
-
-    def encode_text(self, caption_ids: np.ndarray) -> np.ndarray:
-        """Encode text captions into a latent space."""
-        logger.info(f"Encoding text with shape: {caption_ids.shape}")
-        caption_ids_tensor = torch.tensor(caption_ids, device=self.device)
-        caption_latents_tensor = self.text_encoder(caption_ids_tensor)[0]
-        logger.info(f"Caption latents shape: {caption_latents_tensor.shape}")
-        return convert_tensor_to_array(caption_latents_tensor)
-
-    def __call__(self, batch: dict[str, np.ndarray]) -> dict[str, np.ndarray]:
-        """Encode images and text captions."""
-        with torch.no_grad():
-            # Step 1: Encode images.
-            input_images = batch[f"image_{self.resolution}"]
-            image_latents = self.encode_images(input_images)
-            # Shape = [batch_size, 4, 32, 32]
-            batch[self.image_latents_key] = image_latents
-
-            del batch[f"image_{self.resolution}"]
-            gc.collect()
-
-            # Step 2: Encode captions.
-            caption_ids = batch["caption_ids"]
-            # Shape = [batch_size, 77, 1024]
-            batch[CAPTION_LATENTS_KEY] = self.encode_text(caption_ids)
-
-            del batch["caption_ids"]
-            gc.collect()
-
-        return batch
 
 
 ########################################
@@ -450,7 +211,7 @@ small_unet_model_config = {
     "attention_head_dim": 8,
     "block_out_channels": [160, 320, 640, 640],
     "center_input_sample": False,
-    "cross_attention_dim": 1024,
+    "cross_attention_dim": 768,
     "down_block_types": [
         "CrossAttnDownBlock2D",
         "CrossAttnDownBlock2D",
@@ -656,7 +417,7 @@ class RayTrainReportCallback(pl.callbacks.Callback):
         # Create a local temporary directory to save the checkpoint.
         temp_checkpoint_dir = os.path.join(
             tempfile.gettempdir(),
-            ray.train.get_context().get_trial_name(),
+            ray.train.get_context().get_experiment_name() or "default",
             f"step={step}",
         )
         os.makedirs(temp_checkpoint_dir, exist_ok=True)
@@ -717,7 +478,7 @@ def train_func(config: dict) -> None:
     """Training function for Stable Diffusion model."""
     seed = config["seed"]
     seed_everything(seed)
-    trial_name = ray.train.get_context().get_trial_name()
+    trial_name = ray.train.get_context().get_experiment_name() or "default"
 
     # Prepare Ray datasets.
     collate_fn = partial(move_to_device_collate_fn, device=ray.train.torch.get_device())
@@ -852,7 +613,7 @@ def train(
     experiment_name: str = "sd-demo",
     seed: int = 2024,
     init_from_pretrained: bool = False,
-    model_name: str = "stabilityai/stable-diffusion-2-base",
+    model_name: str = "CompVis/stable-diffusion-v1-4",
     use_xformers: bool = False,
     batch_size_per_worker: int = 1,
     max_steps: int = 20,
@@ -885,8 +646,8 @@ def train(
     validation_data_uri: str = "s3://anyscale-materials/stable-diffusion/laion_art_sample_valid.parquet",
 ):
     """Train a Stable Diffusion model."""
-    ray.data.set_progress_bars(False)
     ctx = ray.data.DataContext.get_current()
+    ctx.enable_progress_bars = False
     ctx.execution_options.verbose_progress = False
 
     train_ds = get_laion_streaming_dataset(
